Modify: get 분리
