--- .vscode/settings.json
@@ -1,3 +1,0 @@
-{
-    "python.pythonPath": "/bin/python"
-}
