+++ seolleungbeer/urls.py
@@ -18,6 +18,6 @@ from django.urls import path, include
 
 urlpatterns = [
     path('users', include('users.urls')),
-    path('products',include('products.urls')),
-    path('orders',include('orders.urls'))
+    path('products', include('products.urls')),
+    path('orders', include('orders.urls'))
 ]
