+++ users/urls.py
@@ -1,10 +1,11 @@
 from django.urls import path
 
-from .views      import SignupView, LoginView, EmailCheckView, MobileCheckView 
+from .views      import SignupView, LoginView, EmailCheckView, MobileCheckView,AccountInfoView
 
 urlpatterns = [
     path('/signup', SignupView.as_view()),
     path('/login', LoginView.as_view()),
     path('/email-check', EmailCheckView.as_view()),
     path('/mobile-check', MobileCheckView.as_view()),
+    path('/account',AccountInfoView.as_view())
 ]
